fix format

+++ examples/textual_inversion/textual_inversion.py
@@ -565,9 +565,7 @@ def main():
                 noisy_latents = noise_scheduler.add_noise(latents, noise, timesteps)
 
                 # Get the text embedding for conditioning
-                encoder_hidden_states = (
-                    text_encoder(batch["input_ids"])[0].to(dtype=weight_dtype).to(dtype=weight_dtype)
-                )
+                encoder_hidden_states = text_encoder(batch["input_ids"])[0].to(dtype=weight_dtype)
 
                 # Predict the noise residual
                 model_pred = unet(noisy_latents, timesteps, encoder_hidden_states).sample
